read/write to Config space should be safe
Move the resposibility for accessing config space safely to the
implementation of the config space. This limits the unsafe boundary and
makes reasoning about safe access easier than providing a thin wrapper
around a raw pointer access and relying on the calling functions to stay
within bounds.

This also requires a mutable reference to self for modifications to the
config space which more accurately reflects that situation.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>